Improve script

+++ app/scripts/fetch_idl.bash
@@ -5,18 +5,20 @@ if [[ -f "$ENVFILE" ]]; then
   cat .env.local
 fi
 
-which cargo
 
 #npx @project-serum/anchor-cli --provider.cluster https://api.mainnet-beta.solana.com idl fetch ${NEXT_PUBLIC_PROGRAM_ADDRESS} > ./idl.json
 #
 
-if [[ "$(uname -a)" == "Linux" ]]; then
+if [[ "$(uname -s)" == "Linux" ]]; then
 apt-get update && apt-get upgrade && apt-get install -y pkg-config build-essential libudev-dev libssl-dev
 fi
 
-cargo install --git https://github.com/coral-xyz/anchor avm --locked --force
+if [[ -x "which -s anchor" ]]; then
+else
+  cargo install --git https://github.com/coral-xyz/anchor avm --locked --force
 
-avm install latest
-avm use latest
+  avm install latest
+  avm use latest
 
-anchor --version
+  anchor --version
+fi
